📝 docs(README.md): update installation instructions for running the app using Docker
The previous instructions for cloning the repository and building the image have been replaced with instructions to pull the image from the GitHub Container Registry. This simplifies the installation process and ensures that the latest version of the image is used.

The command to run the container has also been updated to remove the unnecessary environment variable `SPOTIPY_REDIRECT_URI` and port mapping `-p 8084:8084`.